Move KeyDelay.h to the editor
