+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/KeyboardManagerEditorLibrary.vcxproj
@@ -40,6 +40,7 @@
     <ClInclude Include="EditShortcutsWindow.h" />
     <ClInclude Include="KeyboardManagerEditorStrings.h" />
     <ClInclude Include="KeyboardManagerState.h" />
+    <ClInclude Include="KeyDelay.h" />
     <ClInclude Include="KeyDropDownControl.h" />
     <ClInclude Include="LoadingAndSavingRemappingHelper.h" />
     <ClInclude Include="pch.h" />
@@ -60,6 +61,7 @@
     <ClCompile Include="EditShortcutsWindow.cpp" />
     <ClCompile Include="KeyboardManagerEditorStrings.cpp" />
     <ClCompile Include="KeyboardManagerState.cpp" />
+    <ClCompile Include="KeyDelay.cpp" />
     <ClCompile Include="KeyDropDownControl.cpp" />
     <ClCompile Include="LoadingAndSavingRemappingHelper.cpp" />
     <ClCompile Include="pch.cpp">
