+++ src/modules/keyboardmanager/common/KeyboardManagerCommon.vcxproj
@@ -47,7 +47,6 @@
     <ClCompile Include="Helpers.cpp" />
     <ClCompile Include="KeyboardEventHandlers.cpp" />
     <ClCompile Include="MappingConfiguration.cpp" />
-    <ClCompile Include="KeyDelay.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(CIBuild)'!='true'">Create</PrecompiledHeader>
     </ClCompile>
@@ -61,7 +60,6 @@
     <ClInclude Include="InputInterface.h" />
     <ClInclude Include="Helpers.h" />
     <ClInclude Include="KeyboardManagerConstants.h" />
-    <ClInclude Include="KeyDelay.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="RemapShortcut.h" />
     <ClInclude Include="Shortcut.h" />
